+++ js/react/bundle.jsx
@@ -26,17 +26,17 @@ window.WishCardsCarousel = WishCardsCarousel;
 import Admin from './pages/admin/index.jsx';
 window.Admin = Admin;
 
-import WishCardCreate from './pages/wishcard/WishCardCreate.jsx';
+import WishCardCreate from './pages/wishcard/create.jsx';
 window.WishCardCreate = WishCardCreate;
 
-import WishCardManage from './pages/wishcard/WishCardManage.jsx';
+import WishCardManage from './pages/wishcard/manage.jsx';
 window.WishCardManage = WishCardManage;
 
 import WishcardsAdministration from './pages/admin/routes/wishcard/administration.jsx';
 window.WishcardsAdministration = WishcardsAdministration;
 
-import AgencyRegister from './pages/signup/AgencyRegister.jsx';
+import AgencyRegister from './pages/profile/agency/register.jsx';
 window.AgencyRegister = AgencyRegister;
 
-import DonationHistory from './pages/profile/DonationHistory.jsx';
+import DonationHistory from './pages/profile/donationhistory.jsx';
 window.DonationHistory = DonationHistory;
